+++ macros/sql/get_tables_by_pattern_sql.sql
@@ -16,6 +16,32 @@
 
 {% endmacro %}
 
+{% macro redshift__get_tables_by_pattern_sql(schema_pattern, table_pattern, exclude='', database=target.database) %}
+
+    {% set sql %}
+        select distinct
+            table_schema as {{ adapter.quote('table_schema') }},
+            table_name as {{ adapter.quote('table_name') }},
+            {{ dbt_utils.get_table_types_sql() }}
+        from "{{ database }}"."information_schema"."tables"
+        where table_schema ilike '{{ schema_pattern }}'
+        and table_name ilike '{{ table_pattern }}'
+        and table_name not ilike '{{ exclude }}'
+        union all
+        select distinct
+            schemaname as {{ adapter.quote('table_schema') }},
+            tablename as {{ adapter.quote('table_name') }},
+            'external' as {{ adapter.quote('table_type') }}
+        from svv_external_tables
+        where redshift_database_name = '{{ database }}'
+        and schemaname ilike '{{ schema_pattern }}'
+        and table_name ilike '{{ table_pattern }}'
+        and table_name not ilike '{{ exclude }}'
+    {% endset %}
+
+    {{ return(sql) }}
+{% endmacro %}
+
 
 {% macro bigquery__get_tables_by_pattern_sql(schema_pattern, table_pattern, exclude='', database=target.database) %}