+++ CHANGELOG.md
@@ -16,13 +16,18 @@
 - deduplicate macro for Databricks now uses the QUALIFY clause, which fixes NULL columns issues from the default natural join logic
 - deduplicate macro for Redshift now uses the QUALIFY clause, which fixes NULL columns issues from the default natural join logic
 - Equality test will now raise an error when the second model has less columns than the first ([#785](https://github.com/dbt-labs/dbt-utils/issues/785))
+- get_tables_by_pattern_sql will now:
+  - return redshift external tables ([#752](https://github.com/dbt-labs/dbt-utils/issues/752)
+  - work with valid redshift database names that contain dashes
+## Under the hood
+- created a new dispatch redshift__get_tables_by_pattern which unions the result of the default macro and querying svv_external_tables for the same conditions (schema name, pattern, exclude pattern).
 
 ## Contributors:
 - [@graciegoheen](https://github.com/graciegoheen)
 - [@yauhen-sobaleu](https://github.com/yauhen-sobaleu)
 - [@rlh1994](https://github.com/rlh1994)
 - [@brunocostalopes](https://github.com/brunocostalopes)
-
+[@brendan-cook-87](https://github.com/brendan-cook-87)
 
 # dbt utils v1.1.1
 ## New features